refactor folder structure
* move component <Code /> to src/components/common
* rename component file name
