+++ packages/react-bootstrap-table2-example/examples/columns/column-format-with-extra-data-table.js
@@ -2,7 +2,7 @@
 import React from 'react';
 
 import { BootstrapTableful } from 'react-bootstrap-table2';
-import Code from 'common/codeBlock';
+import Code from 'components/common/code-block';
 import { productsGenerator } from 'utils/common';
 
 const products = productsGenerator(5, (value, index) => ({
