+++ packages/react-bootstrap-table2-example/src/components/common/code-block.js
